fix: move datadog into utils
