+++ src/bootstrap.js
@@ -2,9 +2,9 @@ import ms from 'ms';
 
 import * as algolia from './algolia/index.js';
 import { config } from './config';
-import { datadog } from './datadog';
 import * as npm from './npm/index.js';
 import saveDocs from './saveDocs.js';
+import { datadog } from './utils/datadog';
 import { log } from './utils/log';
 
 let loopStart;
