+++ src/index.js
@@ -2,8 +2,8 @@ import * as algolia from './algolia/index.js';
 import * as bootstrap from './bootstrap.js';
 import { config } from './config';
 import createStateManager from './createStateManager.js';
-import { datadog } from './datadog';
 import * as jsDelivr from './jsDelivr/index.js';
+import { datadog } from './utils/datadog';
 import { log } from './utils/log';
 import * as sentry from './utils/sentry';
 import * as watch from './watch.js';
