+++ src/npm/index.js
@@ -3,7 +3,7 @@ import nano from 'nano';
 import numeral from 'numeral';
 
 import { config } from '../config';
-import { datadog } from '../datadog';
+import { datadog } from '../utils/datadog';
 import { log } from '../utils/log';
 import { request } from '../utils/request';
