+++ src/typescriptSupport.js
@@ -1,6 +1,6 @@
-import { datadog } from './datadog';
 import * as npm from './npm/index.js';
 import { fileExistsInUnpkg } from './unpkg.js';
+import { datadog } from './utils/datadog';
 
 /**
  * Basically either
